MemoryTests: Make time-sensitive test to run in serial order.
